+++ libs/ng-bpmn/src/index.ts
@@ -5,5 +5,6 @@ export * from './lib/ng-cmmn/ng-cmmn.component';
 export * from './lib/core/ImportEvent';
 export * from './lib/core/Modeler';
 export * from './lib/core/modeling/ModelerActions';
+export * from './lib/core/modeling/DiagramComments';
 
 export * from './lib/services/ModelingService';