+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- ng-bpmn: `showComments` property to enable embedded comments via [bpmn-js-embedded-comments](https://github.com/bpmn-io/bpmn-js-embedded-comments)
 - new `ModelingService` to handle frequently used modeling actions (i.e. export to SVG/XML, etc.)
 - app: new toolbar actions: Undo/Redo, Zoom